Foliage/Drops, DayNight Cycle and Seasons Working
Foliage is now generating and being removed when the player interacts with the foliage, so the player can receive logs or other drops from interacting with the foliage. Day Night Cycle is set up to work for the course of a year at a time (day, month, year tracking) and seasons change depending on the month and position relative to equator (Northern Hemisphere, Southern Hemisphere).